Fixed minor compiler warnings.

+++ include/coap-te/message/impl/serialize.ipp
@@ -32,7 +32,7 @@ serialize_header(type tp, code co, message_id mid,
   static_assert(coap_te::core::is_const_buffer_type_v<ConstBuffer>, "Must be const buffer type");
   static_assert(coap_te::core::is_mutable_buffer_type_v<MutableBuffer>, "Must be mutable buffer type");
 
-  std::size_t token_len = (std::min)(token.size(), max_token_size);
+  std::uint8_t token_len = static_cast<std::uint8_t>((std::min)(token.size(), max_token_size));
   std::size_t size = minimum_header_size + token_len;
   if (output.size() < size) {
     ec = std::make_error_code(std::errc::no_buffer_space);
